fix(riscv/sbi): do not truncate the HSM hart_start opaque (priv)
The SBI HSM hart_start 'opaque' argument is a full xlen value (Linux passes a
64-bit guest-physical boot-data pointer), but it was stored/read as 'unsigned',
truncating it to 32 bits and faulting the secondary harts. Use unsigned long.

+++ src/arch/riscv/sbi.c
@@ -351,7 +351,7 @@ static struct sbiret sbi_hsm_start_handler(void)
                 ret.error = SBI_ERR_FAILURE;
             } else {
                 vaddr_t start_addr = vcpu_readreg(cpu()->vcpu, REG_A1);
-                unsigned priv = (unsigned)vcpu_readreg(cpu()->vcpu, REG_A2);
+                unsigned long priv = vcpu_readreg(cpu()->vcpu, REG_A2);
                 vcpu->arch.sbi_ctx.state = START_PENDING;
                 vcpu->arch.sbi_ctx.start_addr = start_addr;
                 vcpu->arch.sbi_ctx.priv = priv;
